Added various improvement to dragon riding
Dragons can be commanded to sit with a stick without the player needing to crouch.
Added option to prevent dismounting a dragon during flight.
If in-flight dismounting is prevented, crouch key can also be used to fly downwards.
Dragons will not wander for 15-30 seconds after their rider dismounted.
Implemented player dismounting to the side of a dragon (used code of vanilla horses).